rename: 스피너 관련 폴더 네이밍 변경
