+++ packages/wow-ui/src/assets/lotties/blueSpinner.json
