+++ packages/wow-ui/src/assets/lotties/rainbowSpinner.json
